Move away from shared server project
Just not a lot that we can share here.  Moving to separate projects for
now.  Will recreate a shared piece later if it makes sense.
