+++ Compilers.sln
@@ -103,8 +103,6 @@ Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "BasicCompilerPerformanceTes
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Server", "Server", "{E35DA3D1-16C0-4318-9187-6B664F12A870}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServerCore", "src\Compilers\Server\Core\ServerCore.csproj", "{9929E5AE-85CA-46A8-A500-3BE7BC634D87}"
-EndProject
 Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "CommandLine", "src\Compilers\Core\CommandLine\CommandLine.shproj", "{AD6F474E-E6D4-4217-91F3-B7AF1BE31CCC}"
 EndProject
 Global
@@ -115,8 +113,8 @@ Global
 		src\Compilers\Core\CommandLine\CommandLine.projitems*{ad6f474e-e6d4-4217-91f3-b7af1be31ccc}*SharedItemsImports = 13
 		src\Compilers\Core\AnalyzerDriver\AnalyzerDriver.projitems*{d0bc9be7-24f6-40ca-8dc6-fcb93bd44b34}*SharedItemsImports = 13
 		src\Compilers\Core\CommandLine\CommandLine.projitems*{e58ee9d7-1239-4961-a0c1-f9ec3952c4c1}*SharedItemsImports = 4
+		src\Compilers\Core\CommandLine\CommandLine.projitems*{9508f118-f62e-4c16-a6f4-7c3b56e166ad}*SharedItemsImports = 4
 		src\Compilers\Core\SharedCollections\SharedCollections.projitems*{afde6bea-5038-4a4a-a88e-dbd2e4088eed}*SharedItemsImports = 4
-		src\Compilers\Core\CommandLine\CommandLine.projitems*{9929e5ae-85ca-46a8-a500-3be7bc634d87}*SharedItemsImports = 4
 		src\Compilers\Core\SharedCollections\SharedCollections.projitems*{1ee8cad3-55f9-4d91-96b2-084641da9a6c}*SharedItemsImports = 4
 		src\Compilers\Core\AnalyzerDriver\AnalyzerDriver.projitems*{1ee8cad3-55f9-4d91-96b2-084641da9a6c}*SharedItemsImports = 4
 		src\Compilers\CSharp\CSharpAnalyzerDriver\CSharpAnalyzerDriver.projitems*{b501a547-c911-4a05-ac6e-274a50dff30e}*SharedItemsImports = 4
@@ -882,26 +880,6 @@ Global
 		{0835A7FB-C472-4F7D-A500-F9FD072CFF21}.Release|x64.Build.0 = Release|Any CPU
 		{0835A7FB-C472-4F7D-A500-F9FD072CFF21}.Release|x86.ActiveCfg = Release|Any CPU
 		{0835A7FB-C472-4F7D-A500-F9FD072CFF21}.Release|x86.Build.0 = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|ARM.ActiveCfg = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|ARM.Build.0 = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|x64.Build.0 = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Debug|x86.Build.0 = Debug|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|ARM.ActiveCfg = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|ARM.Build.0 = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|Mixed Platforms.Build.0 = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|x64.ActiveCfg = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|x64.Build.0 = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|x86.ActiveCfg = Release|Any CPU
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -948,7 +926,6 @@ Global
 		{ECBB1F28-2672-4344-8CB3-83882F291739} = {32A48625-F0AD-419D-828B-A50BDABA38EA}
 		{1DFEA9C5-973C-4179-9B1B-0F32288E1EF2} = {A41D1B99-F489-4C43-BBDF-96D61B19A6B9}
 		{0835A7FB-C472-4F7D-A500-F9FD072CFF21} = {C65C6143-BED3-46E6-869E-9F0BE6E84C37}
-		{9929E5AE-85CA-46A8-A500-3BE7BC634D87} = {E35DA3D1-16C0-4318-9187-6B664F12A870}
 		{AD6F474E-E6D4-4217-91F3-B7AF1BE31CCC} = {A41D1B99-F489-4C43-BBDF-96D61B19A6B9}
 	EndGlobalSection
 EndGlobal
