--- src/Compilers/Server/Core/project.json
@@ -1,12 +1,0 @@
-{
-  "dependencies": { },
-  "frameworks": {
-    ".NETPortable,Version=v4.5,Profile=Profile7": {}
-  },
-  "runtimes": {
-    "win7": { },
-    "win7-anycpu": { },
-    "ubuntu.14.04": { },
-    "ubuntu.14.04-anycpu": { }
-  }
-}
